+++ src/main/java/insane96mcp/survivalreimagined/module/experience/enchantment/WaterCoolant.java
@@ -1,0 +1,58 @@
+package insane96mcp.survivalreimagined.module.experience.enchantment;
+
+import insane96mcp.survivalreimagined.module.experience.feature.EnchantmentsFeature;
+import insane96mcp.survivalreimagined.utils.Utils;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.item.AxeItem;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.enchantment.DamageEnchantment;
+import net.minecraft.world.item.enchantment.Enchantment;
+import net.minecraft.world.item.enchantment.EnchantmentCategory;
+
+public class WaterCoolant extends Enchantment {
+    public WaterCoolant() {
+        super(Rarity.UNCOMMON, EnchantmentCategory.WEAPON, new EquipmentSlot[]{EquipmentSlot.MAINHAND});
+    }
+
+    @Override
+    public int getMaxLevel() {
+        return 5;
+    }
+
+    @Override
+    public int getMinCost(int lvl) {
+        return 5 + (lvl - 1) * 8;
+    }
+
+    @Override
+    public int getMaxCost(int lvl) {
+        return this.getMinCost(lvl) + 20;
+    }
+
+    @Override
+    public boolean checkCompatibility(Enchantment enchantment) {
+        return !(enchantment instanceof DamageEnchantment);
+    }
+
+    @Override
+    public boolean canEnchant(ItemStack stack) {
+        return stack.getItem() instanceof AxeItem || super.canEnchant(stack);
+    }
+
+    @Override
+    public boolean canApplyAtEnchantingTable(ItemStack stack) {
+        return super.canApplyAtEnchantingTable(stack);
+    }
+
+    @Override
+    public void doPostAttack(LivingEntity attacker, Entity entity, int lvl) {
+        if (!(entity instanceof LivingEntity livingEntity)
+                || !Utils.isEntityInTag(livingEntity, EnchantmentsFeature.WATER_COOLANT_AFFECTED))
+            return;
+
+        if (lvl > 0)
+            livingEntity.setTicksFrozen(Entity.BASE_TICKS_REQUIRED_TO_FREEZE + 10 + (Entity.FREEZE_HURT_FREQUENCY * 2 * lvl));
+    }
+}